Zusätzliche feine Gewichtung der Areas des GameBoard hinsichtlich der Gewinnmöglichkeit und Implementierung in den Minimax Algorythmus. Der Algorythmus gibt jetzt double-Werte zurück.

+++ Logic/TicTacToe.Core/IAI.cs
@@ -7,8 +7,8 @@ namespace NEUKO.TicTacToe.Core
     public interface IAI
     {
         void ShowAITest();
-        int GetAreaIDForX();
-        int GetAreaIDForO();
+        double GetAreaIDForX();
+        double GetAreaIDForO();
         int AreaIDForX { get; }
         int AreaIDForO { get; }
         int MaximumDepth { get; set; }
